+++ test/nn/conv/test_message_passing.py
@@ -159,6 +159,9 @@ def test_my_conv_edge_index():
 
 class MyCommentedConv(MessagePassing):
     r"""This layer calls `self.propagate()` internally."""
+    def __init__(self) -> None:
+        super().__init__()
+
     def forward(self, x: Tensor, edge_index: Tensor) -> Tensor:
         # `self.propagate()` is used here to propagate messages.
         return self.propagate(edge_index, x=x)
@@ -177,6 +180,9 @@ def test_my_commented_conv():
 
 
 class MyKwargsConv(MessagePassing):
+    def __init__(self) -> None:
+        super().__init__()
+
     def forward(self, x: Tensor, edge_index: Tensor) -> Tensor:
         return self.propagate(x=x, edge_index=edge_index)