+++ test/test_edge_index.py
@@ -7,9 +7,8 @@ import torch
 from torch import Tensor, tensor
 
 import torch_geometric
-from torch_geometric import EdgeIndex
+from torch_geometric import EdgeIndex, Index
 from torch_geometric.edge_index import (
-    SUPPORTED_DTYPES,
     ReduceType,
     SortReturnType,
     _scatter_spmm,
@@ -25,10 +24,10 @@ from torch_geometric.testing import (
     withoutExtensions,
     withPackage,
 )
-from torch_geometric.typing import SparseTensor
+from torch_geometric.typing import INDEX_DTYPES, SparseTensor
 from torch_geometric.utils import scatter
 
-DTYPES = [pytest.param(dtype, id=str(dtype)[6:]) for dtype in SUPPORTED_DTYPES]
+DTYPES = [pytest.param(dtype, id=str(dtype)[6:]) for dtype in INDEX_DTYPES]
 IS_UNDIRECTED = [
     pytest.param(False, id='directed'),
     pytest.param(True, id='undirected'),
@@ -47,12 +46,8 @@ def test_basic(dtype, device):
     adj.validate()
     assert isinstance(adj, EdgeIndex)
 
-    if torch_geometric.typing.WITH_PT112:
-        assert str(adj).startswith('EdgeIndex([[0, 1, 1, 2],\n'
-                                   '           [1, 0, 2, 1]], ')
-    else:
-        assert str(adj).startswith('tensor([[0, 1, 1, 2],\n'
-                                   '        [1, 0, 2, 1]], ')
+    assert str(adj).startswith('EdgeIndex([[0, 1, 1, 2],\n'
+                               '           [1, 0, 2, 1]], ')
     assert 'sparse_size=(3, 3), nnz=4' in str(adj)
     assert (f"device='{device}'" in str(adj)) == adj.is_cuda
     assert (f'dtype={dtype}' in str(adj)) == (dtype != torch.long)
@@ -89,6 +84,7 @@ def test_identity(dtype, device, is_undirected):
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], sparse_size=(3, 3), **kwargs)
 
     out = EdgeIndex(adj)
+    assert not isinstance(out.as_tensor(), EdgeIndex)
     assert out.data_ptr() == adj.data_ptr()
     assert out.dtype == adj.dtype
     assert out.device == adj.device
@@ -141,6 +137,8 @@ def test_set_tuple_item():
 
 
 def test_validate():
+    with pytest.raises(TypeError, match="tensors of a single element"):
+        EdgeIndex([torch.tensor([0, 1]), torch.tensor([1, 0])])
     with pytest.raises(ValueError, match="unsupported data type"):
         EdgeIndex([[0.0, 1.0], [1.0, 0.0]])
     with pytest.raises(ValueError, match="needs to be two-dimensional"):
@@ -150,9 +148,9 @@ def test_validate():
     with pytest.raises(ValueError, match="received a non-symmetric size"):
         EdgeIndex([[0, 1], [1, 0]], is_undirected=True, sparse_size=(2, 3))
     with pytest.raises(TypeError, match="invalid combination of arguments"):
-        EdgeIndex(torch.tensor([[0, 1], [1, 0]]), torch.long)
+        EdgeIndex(tensor([[0, 1], [1, 0]]), torch.long)
     with pytest.raises(TypeError, match="invalid keyword arguments"):
-        EdgeIndex(torch.tensor([[0, 1], [1, 0]]), dtype=torch.long)
+        EdgeIndex(tensor([[0, 1], [1, 0]]), dtype=torch.long)
     with pytest.raises(ValueError, match="contains negative indices"):
         EdgeIndex([[-1, 0], [0, 1]]).validate()
     with pytest.raises(ValueError, match="than its number of rows"):
@@ -305,6 +303,7 @@ def test_to_function(dtype, device, is_undirected):
 def test_cpu_cuda(dtype):
     kwargs = dict(dtype=dtype)
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], **kwargs)
+    assert adj.is_cpu
 
     out = adj.cuda()
     assert isinstance(out, EdgeIndex)
@@ -312,7 +311,7 @@ def test_cpu_cuda(dtype):
 
     out = out.cpu()
     assert isinstance(out, EdgeIndex)
-    assert not out.is_cuda
+    assert out.is_cpu
 
 
 @withCUDA
@@ -328,6 +327,15 @@ def test_share_memory(dtype, device):
     assert out._data.is_shared()
     assert out._indptr.is_shared()
     assert out.data_ptr() == adj.data_ptr()
+
+
+@onlyCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_pin_memory(dtype):
+    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], dtype=dtype)
+    assert not adj.is_pinned()
+    out = adj.pin_memory()
+    assert out.is_pinned()
 
 
 @withCUDA
@@ -530,6 +538,20 @@ def test_getitem(dtype, device, is_undirected):
     assert out.is_sorted_by_row
     assert not out.is_undirected
 
+    out = adj[...]
+    assert isinstance(out, EdgeIndex)
+    assert out.equal(tensor([[0, 1, 1, 2], [1, 0, 2, 1]], device=device))
+    assert out.is_sorted_by_row
+    assert out.is_undirected == is_undirected
+
+    out = adj[None]
+    assert not isinstance(out, EdgeIndex)
+    assert out.equal(tensor([[[0, 1, 1, 2], [1, 0, 2, 1]]], device=device))
+
+    out = adj[0, 0]
+    assert not isinstance(out, EdgeIndex)
+    assert out.equal(tensor(0, device=device))
+
     out = adj[:, 0]
     assert not isinstance(out, EdgeIndex)
 
@@ -538,6 +560,95 @@ def test_getitem(dtype, device, is_undirected):
 
     out = adj[tensor([0], device=device), tensor([0], device=device)]
     assert not isinstance(out, EdgeIndex)
+
+
+@withCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_select(dtype, device):
+    kwargs = dict(dtype=dtype, device=device)
+
+    adj = EdgeIndex(
+        [[0, 1, 1, 2], [1, 0, 2, 1]],
+        sort_order='row',
+        sparse_size=(4, 5),
+        **kwargs,
+    ).fill_cache_()
+
+    out = adj[0]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([0, 1, 1, 2], device=device))
+    assert out.dim_size == 4
+    assert out.is_sorted
+    assert out._indptr.equal(tensor([0, 1, 3, 4, 4], device=device))
+
+    out = adj[-1]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([1, 0, 2, 1], device=device))
+    assert out.dim_size == 5
+    assert not out.is_sorted
+    assert out._indptr is None
+
+    out = adj[-2, 2:4]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([1, 2], device=device))
+    assert out.dim_size == 4
+    assert out.is_sorted
+    assert out._indptr is None
+
+    adj = EdgeIndex(
+        [[1, 0, 2, 1], [0, 1, 1, 2]],
+        sort_order='col',
+        sparse_size=(5, 4),
+        **kwargs,
+    ).fill_cache_()
+
+    out = adj[1]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([0, 1, 1, 2], device=device))
+    assert out.dim_size == 4
+    assert out.is_sorted
+    assert out._indptr.equal(tensor([0, 1, 3, 4, 4], device=device))
+
+    out = adj[-2]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([1, 0, 2, 1], device=device))
+    assert out.dim_size == 5
+    assert not out.is_sorted
+    assert out._indptr is None
+
+    out = adj[-1, 2:4]
+    assert isinstance(out, Index)
+    assert out.equal(tensor([1, 2], device=device))
+    assert out.dim_size == 4
+    assert out.is_sorted
+    assert out._indptr is None
+
+
+@withCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_unbind(dtype, device):
+    kwargs = dict(dtype=dtype, device=device)
+
+    adj = EdgeIndex(
+        [[0, 1, 1, 2], [1, 0, 2, 1]],
+        sort_order='row',
+        sparse_size=(4, 5),
+        **kwargs,
+    ).fill_cache_()
+
+    row, col = adj
+
+    assert isinstance(row, Index)
+    assert row.equal(tensor([0, 1, 1, 2], device=device))
+    assert row.dim_size == 4
+    assert row.is_sorted
+    assert row._indptr.equal(tensor([0, 1, 3, 4, 4], device=device))
+
+    assert isinstance(col, Index)
+    assert col.equal(tensor([1, 0, 2, 1], device=device))
+    assert col.dim_size == 5
+    assert not col.is_sorted
+    assert col._indptr is None
 
 
 @withCUDA
@@ -638,7 +749,6 @@ def test_to_sparse_csr(dtype, device):
 
 @withCUDA
 @pytest.mark.parametrize('dtype', DTYPES)
-@pytest.mark.skipif(not torch_geometric.typing.WITH_PT112, reason="<1.12")
 def test_to_sparse_csc(dtype, device):
     kwargs = dict(dtype=dtype, device=device)
     with pytest.raises(ValueError, match="not sorted"):
@@ -684,19 +794,19 @@ def test_add(dtype, device, is_undirected):
     assert out.is_undirected == is_undirected
     assert out.sparse_size() == (7, 7)
 
-    out = adj + torch.tensor([2], dtype=dtype, device=device)
+    out = adj + tensor([2], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [3, 2, 4, 3]], device=device))
     assert out.is_undirected == is_undirected
     assert out.sparse_size() == (5, 5)
 
-    out = adj + torch.tensor([[2], [1]], dtype=dtype, device=device)
+    out = adj + tensor([[2], [1]], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [2, 1, 3, 2]], device=device))
     assert not out.is_undirected
     assert out.sparse_size() == (5, 4)
 
-    out = adj + torch.tensor([[2], [2]], dtype=dtype, device=device)
+    out = adj + tensor([[2], [2]], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [3, 2, 4, 3]], device=device))
     assert out.is_undirected == is_undirected
@@ -731,19 +841,19 @@ def test_sub(dtype, device, is_undirected):
     assert out.is_undirected == is_undirected
     assert out.sparse_size() == (3, 3)
 
-    out = adj - torch.tensor([2], dtype=dtype, device=device)
+    out = adj - tensor([2], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [3, 2, 4, 3]], device=device))
     assert out.is_undirected == is_undirected
     assert out.sparse_size() == (5, 5)
 
-    out = adj - torch.tensor([[2], [1]], dtype=dtype, device=device)
+    out = adj - tensor([[2], [1]], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [4, 3, 5, 4]], device=device))
     assert not out.is_undirected
     assert out.sparse_size() == (5, 6)
 
-    out = adj - torch.tensor([[2], [2]], dtype=dtype, device=device)
+    out = adj - tensor([[2], [2]], dtype=dtype, device=device)
     assert isinstance(out, EdgeIndex)
     assert out.equal(tensor([[2, 3, 3, 4], [3, 2, 4, 3]], device=device))
     assert out.is_undirected == is_undirected
@@ -1000,42 +1110,71 @@ def test_matmul(without_extensions, device):
 
 
 @withCUDA
-def test_sparse_narrow(device):
-    adj = EdgeIndex(
-        [[0, 1, 1, 2], [1, 0, 2, 1]],
-        device=device,
-        sort_order='row',
-    )
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_sparse_row_narrow(dtype, device):
+    kwargs = dict(dtype=dtype, device=device)
+    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], sort_order='row', **kwargs)
 
     out = adj.sparse_narrow(dim=0, start=1, length=1)
-    assert out.equal(torch.tensor([[0, 0], [0, 2]], device=device))
+    assert out.equal(tensor([[0, 0], [0, 2]], device=device))
     assert out.sparse_size() == (1, None)
     assert out.sort_order == 'row'
-    assert out._indptr.equal(torch.tensor([0, 2], device=device))
+    assert out._indptr.equal(tensor([0, 2], device=device))
 
     out = adj.sparse_narrow(dim=0, start=2, length=0)
-    assert out.equal(torch.tensor([[], []], device=device))
+    assert out.equal(tensor([[], []], device=device))
     assert out.sparse_size() == (0, None)
     assert out.sort_order == 'row'
     assert out._indptr is None
 
     out = adj.sparse_narrow(dim=1, start=1, length=1)
-    assert (out.equal(torch.tensor([[0, 2], [0, 0]], device=device))
-            or out.equal(torch.tensor([[2, 0], [0, 0]], device=device)))
+    assert out.equal(tensor([[0, 2], [0, 0]], device=device))
     assert out.sparse_size() == (3, 1)
-    assert out.sort_order == 'col'
-    assert out._indptr.equal(torch.tensor([0, 2], device=device))
+    assert out.sort_order == 'row'
+    assert out._indptr is None
 
     out = adj.sparse_narrow(dim=1, start=2, length=0)
-    assert out.equal(torch.tensor([[], []], device=device))
+    assert out.equal(tensor([[], []], device=device))
     assert out.sparse_size() == (3, 0)
+    assert out.sort_order == 'row'
+    assert out._indptr is None
+
+
+@withCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_sparse_col_narrow(dtype, device):
+    kwargs = dict(dtype=dtype, device=device)
+    adj = EdgeIndex([[1, 0, 2, 1], [0, 1, 1, 2]], sort_order='col', **kwargs)
+
+    out = adj.sparse_narrow(dim=1, start=1, length=1)
+    assert out.equal(tensor([[0, 2], [0, 0]], device=device))
+    assert out.sparse_size() == (None, 1)
+    assert out.sort_order == 'col'
+    assert out._indptr.equal(tensor([0, 2], device=device))
+
+    out = adj.sparse_narrow(dim=1, start=2, length=0)
+    assert out.equal(tensor([[], []], device=device))
+    assert out.sparse_size() == (None, 0)
+    assert out.sort_order == 'col'
+    assert out._indptr is None
+
+    out = adj.sparse_narrow(dim=0, start=1, length=1)
+    assert out.equal(tensor([[0, 0], [0, 2]], device=device))
+    assert out.sparse_size() == (1, 3)
+    assert out.sort_order == 'col'
+    assert out._indptr is None
+
+    out = adj.sparse_narrow(dim=0, start=2, length=0)
+    assert out.equal(tensor([[], []], device=device))
+    assert out.sparse_size() == (0, 3)
     assert out.sort_order == 'col'
     assert out._indptr is None
 
 
 @withCUDA
-def test_sparse_resize(device):
-    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], device=device)
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_sparse_resize(dtype, device):
+    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], dtype=dtype, device=device)
 
     out = adj.sort_by('row')[0].fill_cache_()
     assert out.sparse_size() == (3, 3)
@@ -1047,6 +1186,10 @@ def test_sparse_resize(device):
     assert out._T_indptr.equal(tensor([0, 1, 3, 4, 4, 4], device=device))
     out = out.sparse_resize_(3, 3)
     assert out.sparse_size() == (3, 3)
+    assert out._indptr.equal(tensor([0, 1, 3, 4], device=device))
+    assert out._T_indptr.equal(tensor([0, 1, 3, 4], device=device))
+    out = out.sparse_resize_(None, None)
+    assert out.sparse_size() == (None, None)
     assert out._indptr is None
     assert out._T_indptr is None
 
@@ -1060,6 +1203,10 @@ def test_sparse_resize(device):
     assert out._T_indptr.equal(tensor([0, 1, 3, 4, 4], device=device))
     out = out.sparse_resize_(3, 3)
     assert out.sparse_size() == (3, 3)
+    assert out._indptr.equal(tensor([0, 1, 3, 4], device=device))
+    assert out._T_indptr.equal(tensor([0, 1, 3, 4], device=device))
+    out = out.sparse_resize_(None, None)
+    assert out.sparse_size() == (None, None)
     assert out._indptr is None
     assert out._T_indptr is None
 
@@ -1080,12 +1227,21 @@ def test_numpy():
 @pytest.mark.parametrize('dtype', DTYPES)
 def test_global_mapping(device, dtype):
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], device=device, dtype=dtype)
-    n_id = torch.tensor([10, 20, 30], device=device, dtype=dtype)
+    n_id = tensor([10, 20, 30], device=device, dtype=dtype)
 
     expected = tensor([[10, 20, 20, 30], [20, 10, 30, 20]], device=device)
     out = n_id[adj]
     assert not isinstance(out, EdgeIndex)
     assert out.equal(expected)
+
+
+@withCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_to_vector(device, dtype):
+    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], device=device, dtype=dtype)
+    out = adj.to_vector()
+    assert not isinstance(out, EdgeIndex)
+    assert out.equal(tensor([1, 3, 5, 7], device=device))
 
 
 @withCUDA
@@ -1114,7 +1270,8 @@ def _collate_fn(edge_indices: List[EdgeIndex]) -> List[EdgeIndex]:
 
 @pytest.mark.parametrize('dtype', DTYPES)
 @pytest.mark.parametrize('num_workers', [0, 2])
-def test_data_loader(dtype, num_workers):
+@pytest.mark.parametrize('pin_memory', [False, True])
+def test_data_loader(dtype, num_workers, pin_memory):
     kwargs = dict(dtype=dtype)
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], sort_order='row', **kwargs)
     adj.fill_cache_()
@@ -1124,6 +1281,7 @@ def test_data_loader(dtype, num_workers):
         batch_size=2,
         num_workers=num_workers,
         collate_fn=_collate_fn,
+        pin_memory=pin_memory,
         drop_last=True,
     )
 
@@ -1133,10 +1291,9 @@ def test_data_loader(dtype, num_workers):
         assert len(batch) == 2
         for adj in batch:
             assert isinstance(adj, EdgeIndex)
-            assert adj.dtype == adj.dtype
-            assert adj.is_shared() == (num_workers > 0)
-            assert adj._data.is_shared() == (num_workers > 0)
-            assert adj._indptr.is_shared() == (num_workers > 0)
+            assert adj.dtype == dtype
+            assert adj.is_shared() != (num_workers == 0) or pin_memory
+            assert adj._data.is_shared() != (num_workers == 0) or pin_memory
 
 
 def test_torch_script():