+++ test/utils/test_segment.py
@@ -4,10 +4,10 @@ import pytest
 import torch
 
 import torch_geometric.typing
+from torch_geometric.index import index2ptr
 from torch_geometric.profile import benchmark
 from torch_geometric.testing import withCUDA, withoutExtensions
 from torch_geometric.utils import scatter, segment
-from torch_geometric.utils.sparse import index2ptr
 
 
 @withCUDA